Include node config instead of iterating over it
In provsioning client.rb seems to be layed down with
perms set to 0600 we need this to be 0644 which seems
standard from looking at a few non-provisioning installs
and the chef-client cookbook. I've inlcuded a client.rb
recipe to manage this correctly since our builds will
fail if this isn't readable by dbuild.