Support versions of Ubuntu greater than 14.04 from PackageCloud.
Currently Ubuntu releases greater than 14.04 are not published to
PackageCloud and consequently we see an error like the following:

> Chef::Exceptions::Package: No candidate version available for chefdk

This change swaps out the `packagecloud` cookbook for the
`chef-ingredient` cookbook, and this allows us to set attributes that
are consumed by the underlying `apt-chef` cookbook. As both chefdk and
delivery-cli packages are supported in `chef-ingredient`, this seemed
like a good, modern, and consistent approach.

For Ubuntu releases greater than 14.04 (currently: 15.04 and 15.10), the
packages from the "trusty" release (14.04) will be used rather than the
"lucid" default which is set in the `apt-chef` cookbook presently
(chefdk and delivery-cli don't have lucid-build packages present on
PackageCloud).

**Note** that if current Ubuntu release packages start to be published
on PackageCloud, the current fix/shim/hack in the `repo` recipe should
short-circuit and thus not get in the way (it's intended to be
future-proof'd).

Finally, for backwards compatability, the value for the attribute
`node['delivery_build']['repo_name']` is preserved meaning that values
of:

* `chef/stable`
* `stable`

are both mapped to the `channel :stable` attribute and values of:

* `chef/current`
* `current`

are both mapped to the `channel :current` attribute of the
`chef_ingredient` resources.