Delivery helper Gemrc.to_yaml with symbols
The `~/.gemrc` is sensitive to `symbols` and `strings`. By default chef
automatically convert attributes that are symbols to strings cousing
the creation of the gemrc file in a wrong way.

Added a new DeliveryHelper::Gemrc.to_yaml that will detect which are the
symbols and it will convert them automatically.

Added extra tests. :smile: