Render `.gemrc` file at compile time
This cookbook has some `chef_gem` resources that tries to install gems
on the build-nodes. If we do not render the `.gemrc` file at compile
time we will fail trying to pull from the wrong configuration.

We need to first of all, render the file and then, install the gems.